requirements/development.txt
-r base.txt
black==25.1.0
ruff==0.12.7
pytest==8.4.1
pytest-xdist==3.8.0
pytest-django==4.11.1
pytest-cov==6.2.1
coverage==7.10.1
